Resized columns in the help text to fit longer property names.
The names of some of our newer properties had become so long that
they were no longer fitting in the column. This caused
https://github.com/thaljef/synopsys-detect-bash-completion to break
because it parses property names out of the help text.

Since the property name field is now 18 chars wider, I also reduced
the property description field by the same amount, so the total
width remains the same.

Ideally, the width of the column should by dyanmically computed
based on the longest property name. But that's more work than
I have time for right now.